convert emit module into emitter
simplifies the usage of emitting things in the processor
